+++ src/Terrabuild/Contracts/BuildProgress.fs
@@ -10,6 +10,7 @@ type IBuildProgress =
 
     abstract TaskScheduled: taskId:string -> label:string -> unit
     abstract BatchScheduled: (string * string) list -> unit
+    abstract TaskWaitingForLock: taskId:string -> unit
     abstract TaskDownloading: taskId:string -> unit
     abstract TaskBuilding: taskId:string -> unit
     abstract TaskUploading: taskId:string -> unit